+++ app/src/App.tsx
@@ -14,9 +14,12 @@ import React from 'react';
 
 import { AngularMaterialPreview } from './components/AngularMaterialPreview';
 import { Footer } from './components/Footer';
+import { ExampleSchemaService } from './core/schemaService';
 
+const schemaService = new ExampleSchemaService();
 export const App = () => (
   <JsonFormsEditor
+    schemaService={schemaService}
     schemaProviders={defaultSchemaProviders}
     schemaDecorators={defaultSchemaDecorators}
     editorTabs={[
